argon2: move `Memory` to its own module
Also makes `Instance` accept `Memory` as a parameter.

This is in preparation for moving slice handling to `Memory`.
